chore(icons): split loader file
